fix: move PostFundingEventListener out of peson folder
TIS21-8298
