+++ tcs-service/src/main/java/com/transformuk/hee/tis/tcs/service/listener/PostFundingEventListener.java
@@ -1,4 +1,4 @@
-package com.transformuk.hee.tis.tcs.service.listener.person;
+package com.transformuk.hee.tis.tcs.service.listener;
 
 import com.transformuk.hee.tis.tcs.service.event.PostFundingCreatedEvent;
 import com.transformuk.hee.tis.tcs.service.event.PostFundingDeletedEvent;
